% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-api.R
\name{start_api}
\alias{start_api}
\title{Start service}
\usage{
start_api(port = 5018)
}
\arguments{
\item{port}{Port on which to run the API, see [plumber::plumb()].}
}
\description{
Start the API using [plumber::plumb()]. Shortcut for some boilerplate code.
}
